SPAR-42 Resolve all compilation errors

+++ Solver/utilities.cpp
@@ -13,8 +13,8 @@ namespace ses {
 		FILE* fp;
 		fp = fopen(file_name, "w");
 		for (unsigned i = 0; i < size; i++) {
-			float output = x[i];
-			fprintf(fp, "%d, %f\n", i, output);
+			double output = (double)x[i];
+			fprintf(fp, "%d, %lf\n", i, output);
 		}
 		fclose(fp);
 	}
